@@ -35,18 +35,18 @@
     },
     "minimum-stability": "stable",
     "require": {
-        "php": ">=5.4.0",
-        "psr/log": "^1.0"
+        "php": "^7.4|^8.0|^8.1",
+        "psr/log": "^1.0|^2.0"
     },
     "scripts": {
         "test": "phpunit",
-        "sniff": "phpcs -p --standard=PSR2 --runtime-set ignore_errors_on_exit 1 --runtime-set ignore_warnings_on_exit 1 src/ tests/",
-        "format": "phpcbf -p --standard=PSR2 --runtime-set ignore_errors_on_exit 1 --runtime-set ignore_warnings_on_exit 1 src/ tests/"
+        "sniff": "phpcs -p --standard=PSR12 --runtime-set ignore_errors_on_exit 1 --runtime-set ignore_warnings_on_exit 1 src/ tests/",
+        "format": "phpcbf -p --standard=PSR12 --runtime-set ignore_errors_on_exit 1 --runtime-set ignore_warnings_on_exit 1 src/ tests/"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8",
         "codacy/coverage": "dev-master",
-        "squizlabs/php_codesniffer": "^2.7"
+        "phpunit/phpunit": "*",
+        "squizlabs/php_codesniffer": "^2.7|^3.7"
     },
     "config": {
         "sort-packages": true
